Add axis frame
Axis often obfuscates the graph lines, adding a frame by default makes it more readable.

At one point, the auto-positioning should be fixed, so that it takes all axes into account, right now it only checks if the top-right combination overlaps.
An immediate improvement would be to use the regular bottom-left combination, but then the lines in all other axes render *on top of* the frame. So I should check to see how to make the lines always become behind all legends.

Also the option to choose on which corner the axes should be located would be nice to add.