fix top_right scaling
I do not know why this is needed, but for some reason it resets the top-right axis to linear each time when opening plot settings whenever the regular bottom-left axis is combined with top-right. This solves this by just hard-setting this setting each time this combination is loaded. Other combinations seem to behave as expected.