Fix guess headers
This is a super minor bug fix that I encountered while working with some data. But I thought it'd be good practice to not push to main at all, but instead commit every change through a PR.

When it tries to guess headers from a file, it will by default try to find something using at least two whitespaces.

If that fails, it tries to use the same delimiter as used for the data itself. It is in this second heuristic that it was still using import_settings as a dictionary, while it is an object nowadays. This fixes the bug.