set limits
Sets the limits when

- Deleting an item
- Changing an item (e.g. when changing the axis of one item)

The reason behind this is in the first case, if I have one item with axes going from 0 to 12, and another with the limit going from 0 to 4, and then delete that first axis. Then it should reset the limits to fit the graphs that are left instead of keeping it from 0 to 12.

In the second case, the use case if if I change the axis of one of the items. So if I set the x-axis for the top item to "top", then it should set the limits so that it fits nicely for both graphs.