Get failsafe from template
Gets the fail safe style from template instead of using the first available option. The reason for this is that this ensures that the dark style will be loaded in dark mode, and the light style will be loaded in light mode. And the template should always, always work.

When people upgrade to version 1.5, the old style will no longer work so this exception will be triggered. By default it selects the top style in the current main branch, which is adwaita-dark. Giving a weird experience for people using light theme. Hard-coding adwaita will have the same issue, but for people using dark theme. This PR solves this.