Fix item transform
The current build has custom item transformation being broken in many cases.

The problem is that _index actually is used, but it is used through the operation variable in the eval function. Flake doesn't understand that.

I solved it by just renaming the index variable to _index, as it's technically not directly used as a variable and this solves the bug while letting flake shut up about errors.