Fix color cycle
When the user picks a color through the color picker, the color cycle detection doesn't work in specific cases.

The reason behind this is that the color picker outputs the color in non-capitalized numbers. So colors that contain letters (e.g. #1A5FB4 for blue) are not detected as being used if the picker has picked exactly that number manually.

This fixes that selection by enforcing changed colors in the color picker to be upper case.