Fix set_attributes
When looping through the dictionary, it deletes attributes during the iteration. This gives an error as dictionaries (and lists) are not allowed to change size. (This is because they're ordered nowadays).

The solution is to make a list of which attributes to delete, and then itterate through that. (I was hoping for a cleaner solution, but that's also the solution that is suggested pretty much everywhere)