Raise data when no data selected
Raised a toast when no data was selected within selection mode.

Sometimes it happens that the actions seem unresponsive, only for me to find out that I am in highlight mode and no data is within my highlight. This gives the user a pop up to notify them if that is the case.