Set active style from style editor
Adds a shortcut button that allows to set the style that is being edited as custom style.

I was a bit unsure about the exact behaviour and what would be the most intuitive, but it seemed useful to have a shortcut here to quickly set the style from the editor as well. In the current implementation it hides the button when "System style" is set as default. So only if "Use Custom Style" is enabled, it shows this to allow for quick changing of the plot style.

Pressing the switch sets the style directly to active style. Deselecting it sets  the active style to the system prefered style.