Add new "thesis" style
Add the style that I use for my own thesis. Basically it's without grids,  slightly longer ticks on the inside on all borders and a more "standard" like color cycle. I also use the same style as the basis for academic articles.

Felt to me that a style like this was missing somewhat, and for me personally it's convenient to ship these as well.