Rewrite style parser.
The style parser is now much closer to matplotlibs implementation with a
few optimizations and modified to work with GFile. This makes it work
with every GFile (e. g. from resources, or remote locations like
smb://).

Because this now returns a valid RcParams object all values in the
dictionary are their correct types (instead of strings), this required
some minor changes elsewhere in the codebase.

Signed-off-by: Christoph Matthias Kohnen <christoph.kohnen@disroot.org>