Rework Import Settings.
Import Settings is now created using self, the file and optional name
and import_mode. If no name or mode is provided the default logic wll be
used (filename and suffix-based identification).
Depending on the mode the params attribute will be populated. This is
queried from an extension to preferences.
Each file now gets their own ImportSettings instance to accomodate for
possible changes in the future.

Signed-off-by: Christoph Matthias Kohnen <christoph.kohnen@disroot.org>