Use actual system style.
When querying for the system preferred style, return the shipped
version. This also includes removing the checkmark for the used style if
no custom style is set and ensures the style is always available and has
the correct values.
Also rename the functions as we are no longer returning a path bur
rather a GFile represantation.

Signed-off-by: Christoph Matthias Kohnen <christoph.kohnen@disroot.org>