Use GFile for exporting figures.
This is a workaround. The matplotlib implementation doesn't support file
like objects, so creating an in-memory buffer isn't suitable here.
Instead we have to create a temporary file, which should always have a
path and be accessible by pythons inbuilt file functions, point
matplotlib to that and move it in the correct place.

Signed-off-by: Christoph Matthias Kohnen <christoph.kohnen@disroot.org>