Define more attributes in blueprint.
Move signal handlers as well as a few other things to blueprint.
Moving Signal handlers helps remove the need to reference a widget by
name in code in most cases. This makes the code a bit cleaner in this
regard.
In order to accomplish that, a few light refactors have to be done. This
way the code is more consistent.

Signed-off-by: Christoph Matthias Kohnen <christoph.kohnen@disroot.org>