Update changelist
Update the changelist for the 1.6.1 patch with the latest bug fixes. I don't think I will push this release today (although I am not directly opposed), but I thought it's good to start updating the changelist now before I forget some parts.

Decided to add the information on top of the whats new section. Resetting it with major releases.