Tune labelsize chooser
While working on a graph, I found the slider for the title and label font sizes to be a bit to coarse, being only able to be set to multiples of 10 percentage points.

This PR sets this to multiple of 5 percentage points. Also allows to set the sizes to be 50%. But I am open to keeping 100% as the minimum.